Chore: 벡터 생성자 및 setter 분리
